projects/requirements.txt
appnope==0.1.0
beautifulsoup4==4.6.0
boto==2.47.0
bz2file==0.98
certifi==2017.4.17
chardet==3.0.3
click==6.7
decorator==4.0.11
Flask==0.12.2
gensim==2.1.0
idna==2.5
ipython==6.1.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
MarkupSafe==1.0
mecab-python3==0.7
mysql-connector==2.1.4
numpy==1.12.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
PyYAML==3.12
requests==2.17.3
scikit-learn==0.18.1
scipy==0.19.0
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
smart-open==1.5.3
SQLAlchemy==1.1.13
traitlets==4.3.2
urllib3==1.21.1
virtualenv==15.1.0
wcwidth==0.1.7
Werkzeug==0.12.2
mojimoji==0.0.5
tensorflow==1.3.0
dm-sonnet==1.12
mstranslator==0.3.2
mysql-connector-python-rf==2.2.2
